feat(subgraphs): set initial configuration for graphql code generation
This allows to use build-guild's graphql-codegen cli tool.

Codegen config will be generated at script run, if the env
variables are not defined, config will be built just to skip
code generation.

Also, added env variables required by this feature, and ignored the
generated gql files as they don't have to be versioned.

Finally, and IMPORTANT: I needed to override a stylelint dependency
as it failed to execute on the pre-commit hook.